feat(firmware): integrate 95 commits from fix/stable-effect-ids branch
Resolves branch divergence where 95 commits (444 files, 35,518 lines)
on fix/stable-effect-ids were never merged to main before the
firmware/v2 -> firmware-v3 restructure.

Strategy: overlay merge — copy complete file state from feature branch
into firmware-v3/, then re-apply main-only additions (MabuTrace
instrumentation, SPDX headers, Trace.h wrapper).

Key integrations:
- ESV11 audio backend (Emotiscope v1.1 Goertzel DSP, 12.8kHz + 32kHz)
- PipelineCore audio backend (256-bin FFT, BeatTracker, FrequencyMap)
- Stable effect ID registry (uint8_t -> EffectId uint16_t, family namespaces)
- RMT driver crash fix (FASTLED_RMT_BUILTIN_DRIVER=0, mutex + interval guard)
- 20+ new effects (BeatPulseBloom, WaterCaustics, StressGlass, etc.)
- Zone preset persistence (EffectPresetManager, ZonePresetManager)
- Serial command module refactor
- Custom board definition (ESP32-S3-DevKitC-1-N16R8)
- No-OTA partition table for development builds
- MabuTrace re-instrumented across AudioActor, RendererActor, LedDriver, main

Build verified:
- esp32dev_audio:              SUCCESS (RAM 28.6%, Flash 32.7%)
- esp32dev_audio_esv11:        SUCCESS (RAM 29.6%, Flash 32.1%)
- esp32dev_audio_esv11_32khz:  SUCCESS (RAM 29.6%, Flash 32.1%)

Integrated commit range: fix/stable-effect-ids 95 commits ahead of
merge-base e0b3174f (2026-01-30 .. 2026-02-22).

+++ firmware-v3/src/codec/WsPluginsCodec.cpp
@@ -1,5 +1,3 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright 2025-2026 SpectraSynq
 /**
  * @file WsPluginsCodec.cpp
  * @brief WebSocket plugins codec implementation
@@ -133,7 +131,7 @@ void WsPluginsCodec::encodePluginsList(const ::lightwaveos::plugins::PluginManag
     
     // List registered effect IDs
     JsonArray effects = data["effects"].to<JsonArray>();
-    for (uint8_t i = 0; i < plugins::PluginConfig::MAX_EFFECTS; i++) {
+    for (uint16_t i = 0; i < plugins::PluginConfig::MAX_EFFECTS; i++) {
         if (manager.isEffectRegistered(i)) {
             effects.add(i);
         }
